Action attack roll
Attack rolls asociated with actions can now be rolled through the character sheet.
Damage rolls are not implemented yet

+++ RPGTable/Assets/Scripts/GameManager.cs
@@ -108,7 +108,10 @@ public class GameManager : NetworkBehaviour
     {
         if (!IsHost)
         {
-            DeleteCharacterServerRpc(characterID);
+            if (NetworkManager.LocalClientId == characterSheets[characterID].ownerID)
+            {
+                DeleteCharacterServerRpc(characterID);
+            }            
         }
         else
         {
